+++ web/src/app/services/ServiceRouter.js
@@ -8,7 +8,7 @@ import ServiceDetails from './ServiceDetails'
 import ServiceLabelList from './ServiceLabelList'
 import IntegrationKeyList from './IntegrationKeyList'
 import { PageNotFound } from '../error-pages/Errors'
-import ServiceAlerts from './components/ServiceAlerts'
+import ServiceAlerts from './ServiceAlerts'
 import ServiceCreateDialog from './ServiceCreateDialog'
 import HeartbeatMonitorList from './HeartbeatMonitorList'
 import { searchSelector } from '../selectors'
